added Java exceptions class

+++ java-app/src/main/java/com/teamkeys/java_app/scrapedData/repo/MatchEvents.java
@@ -2,5 +2,5 @@ package com.teamkeys.java_app.scrapedData.repo;
 
 public interface MatchEvents {
  int getMatchId();
- long getEventsCount();
+ short getEventsCount();
 }
